# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Maricho API
  description: Backend API for the Maricho skilled work platform.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /jobs/request:
    post:
      summary: Create a Quick Hire job request
      description: >
        Buyer submits a problem description (text, or a voice-note transcript
        arriving as text from the WhatsApp/client channel). The buyer is taken
        from the bearer token, not the request body.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - trade
                - suburb
                - address
                - problem_description
              properties:
                trade:
                  type: string
                  enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
                suburb:
                  type: string
                address:
                  type: string
                  description: >
                    Exact street address. Stored at request time but only
                    ever shown to the buyer, the assigned worker (from
                    booking onward), or OPS — never to unassigned match
                    candidates.
                landmark_narrative:
                  type: string
                  nullable: true
                  description: >-
                    Optional navigation description, gated the same as
                    address. See ADR-004.
                latitude:
                  type: number
                  nullable: true
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  nullable: true
                  minimum: -180
                  maximum: 180
                problem_description:
                  type: string
                problem_photo_url:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Job requested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          description: Missing or invalid bearer token
        '403':
          description: Authenticated user is not a BUYER
        '422':
          description: >-
            Invalid input (missing/blank suburb or description, unknown
            suburb, or out-of-range latitude/longitude)
  /jobs/{job_id}:
    get:
      summary: Fetch a job
      description: >
        Restricted to the job's buyer, its assigned worker, or an OPS user.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Authenticated user is not a party to this job
        '404':
          description: Job not found
  /jobs/{job_id}/matches:
    get:
      summary: Run the matching engine and return ranked worker candidates
      description: >
        Returns up to 3 workers filtered to the job's trade, ranked primarily
        by distance band (SAME_SUBURB > NEARBY > MODERATE; FAR is excluded)
        and secondarily by standing (grade, on-time rate, dispute rate, fill
        rate) within the same band. Distance is computed via haversine
        distance between suburb centroids (see the `suburbs` reference
        table). Excludes workers currently tied to another active job. Each
        candidate includes a `reason` explaining the match. The first call
        on a REQUESTED job transitions it to MATCHED; a job past BOOKED no
        longer accepts matches. Restricted to the job's buyer.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Up to 3 ranked candidates (possibly empty)
          content:
            application/json:
              schema:
                type: array
                maxItems: 3
                items:
                  $ref: '#/components/schemas/MatchCandidate'
        '403':
          description: Authenticated user is not this job's buyer
        '404':
          description: Job not found
        '409':
          description: Job is no longer accepting matches (already BOOKED or later)
  /jobs/{job_id}/book:
    post:
      summary: Book a matched worker
      description: >
        Buyer picks one of the matched workers, holds a deposit
        (DEPOSIT_HELD ledger entry), and reveals the exact address to the
        worker. The chosen worker must be a registered WORKER with a skill
        in this job's trade and not already tied to another active job.
        `currency` (ADR-002, default USD) is set once here and reused by
        every later ledger entry for this job. When `currency` is
        `ECOCASH`, a real collection request is initiated from the buyer
        before the booking completes — see 503/502 below.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_id, deposit_amount]
              properties:
                worker_id:
                  type: string
                  format: uuid
                deposit_amount:
                  type: number
                  exclusiveMinimum: 0
                currency:
                  type: string
                  enum: [USD, ECOCASH, ZWG]
                  default: USD
      responses:
        '200':
          description: Job booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Authenticated user is not this job's buyer
        '404':
          description: Job not found
        '409':
          description: Job is not MATCHED, or the worker is already committed elsewhere
        '422':
          description: worker_id does not refer to a registered, skilled worker
        '502':
          description: currency was ECOCASH and the collection request failed
        '503':
          description: currency was ECOCASH and no gateway config exists yet
  /jobs/{job_id}/quote:
    post:
      summary: Assigned worker submits their on-site price
      description: >
        Two-tier quoting (ADR-003): the worker submits a labor amount and a
        materials amount separately; quote_amount is computed server-side as
        their sum (labor + materials), never accepted directly from the
        client. Commits the remaining balance (quote minus the deposit
        already held) as a BALANCE_COMMITTED ledger entry and moves the job
        to IN_PROGRESS. Only the job's assigned worker may call this.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labor_amount, materials_amount]
              properties:
                labor_amount:
                  type: number
                  exclusiveMinimum: 0
                materials_amount:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Authenticated user is not this job's assigned worker
        '404':
          description: Job not found
        '409':
          description: Job is not BOOKED
        '422':
          description: Quote does not exceed the deposit already held
  /jobs/{job_id}/complete:
    post:
      summary: Buyer signs off and releases payment
      description: >
        Writes the immutable RecordEntry and a RELEASED ledger entry (the
        full quote amount). Only the job's buyer may call this.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [what_was_done]
              properties:
                what_was_done:
                  type: string
                client_words:
                  type: string
                  nullable: true
                before_photo_url:
                  type: string
                  nullable: true
                after_photo_url:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Job completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Authenticated user is not this job's buyer
        '404':
          description: Job not found
        '409':
          description: Job is not IN_PROGRESS
  /jobs/{job_id}/dispute:
    post:
      summary: Buyer raises a dispute
      description: >
        Money freezes: the job moves to DISPUTED and no further ledger
        release can happen until the worker responds. "Design for the
        worker's worst day" — the worker's standing is frozen, not dropped,
        while the dispute is open (see GET /workers/me/standing).
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '201':
          description: Dispute opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '403':
          description: Authenticated user is not this job's buyer
        '404':
          description: Job not found
        '409':
          description: Job is not IN_PROGRESS
  /jobs/{job_id}/dispute/respond:
    post:
      summary: Assigned worker answers the dispute
      description: >
        Default outcome is a return visit, not a refund: the dispute
        resolves to RESOLVED_RETURN_VISIT and the job goes back to
        IN_PROGRESS so the worker can fix the issue and the buyer can
        complete normally afterward. Only the job's assigned worker may
        call this. Photos are attached as part of the response, not
        sourced automatically — there is no separate execution-phase photo
        upload endpoint yet.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: string
                before_photo_url:
                  type: string
                  nullable: true
                after_photo_url:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '403':
          description: Authenticated user is not this job's assigned worker
        '404':
          description: Job not found, or no open dispute
        '409':
          description: Job is not currently DISPUTED
  /jobs/{job_id}/disputes:
    get:
      summary: List a job's disputes
      description: Restricted to the job's buyer, its assigned worker, or an OPS user.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Disputes, oldest first, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dispute'
        '403':
          description: Authenticated user is not a party to this job
        '404':
          description: Job not found
  /jobs/{job_id}/ledger:
    get:
      summary: View a job's ledger entries
      description: Restricted to the job's buyer, its assigned worker, or an OPS user.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Ledger entries, oldest first, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntry'
        '403':
          description: Authenticated user is not a party to this job
        '404':
          description: Job not found
  /jobs/{job_id}/record:
    get:
      summary: View a job's immutable record entry
      description: Restricted to the job's buyer, its assigned worker, or an OPS user.
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Record entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordEntry'
        '403':
          description: Authenticated user is not a party to this job
        '404':
          description: Job not found, or no record entry yet (job not completed)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
      description: Maximum rows to return.
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Rows to skip before collecting the result set.
  schemas:
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [REQUESTED, MATCHED, BOOKED, IN_PROGRESS, DISPUTED, COMPLETED, CANCELLED]
        trade:
          type: string
          enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
        suburb:
          type: string
        address:
          type: string
          description: Only present for the buyer, the assigned worker, or OPS.
        landmark_narrative:
          type: string
          nullable: true
          description: >-
            Optional navigation description (e.g. "blue gate opposite the
            clinic"), gated the same as address.
        latitude:
          type: number
          nullable: true
          description: Optional GPS pin, gated the same as address.
        longitude:
          type: number
          nullable: true
          description: Optional GPS pin, gated the same as address.
        problem_description:
          type: string
          nullable: true
        problem_photo_url:
          type: string
          nullable: true
        currency:
          type: string
          enum: [USD, ECOCASH, ZWG]
          nullable: true
          description: Set once at booking (ADR-002); nullable until then.
        labor_amount:
          type: number
          nullable: true
          description: Set together with materials_amount by POST /jobs/{id}/quote (ADR-003).
        materials_amount:
          type: number
          nullable: true
        quote_amount:
          type: number
          nullable: true
          description: Computed server-side as labor_amount + materials_amount.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        amount:
          type: number
        entry_type:
          type: string
          enum: [DEPOSIT_HELD, BALANCE_COMMITTED, RELEASED, REFUNDED]
        currency:
          type: string
          enum: [USD, ECOCASH, ZWG]
        external_reference:
          type: string
          nullable: true
          description: Gateway transaction reference (ADR-002), set only for ECOCASH entries.
        created_at:
          type: string
          format: date-time
    RecordEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
        what_was_done:
          type: string
        client_words:
          type: string
          nullable: true
        before_photo_url:
          type: string
          nullable: true
        after_photo_url:
          type: string
          nullable: true
        signed_off_at:
          type: string
          format: date-time
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        raised_by_id:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [OPEN, RESOLVED_RETURN_VISIT]
        worker_response:
          type: string
          nullable: true
        before_photo_url:
          type: string
          nullable: true
        after_photo_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        responded_at:
          type: string
          format: date-time
          nullable: true
    MatchCandidate:
      type: object
      properties:
        worker_id:
          type: string
          format: uuid
        suburb:
          type: string
        travel_means:
          type: string
          nullable: true
        distance_band:
          type: string
          enum: [SAME_SUBURB, NEARBY, MODERATE, FAR]
        distance_km:
          type: number
        grade:
          type: string
        on_time_rate:
          type: number
        dispute_rate:
          type: number
        fill_rate:
          type: number
        jobs_completed:
          type: integer
        reason:
          type: string
...